Update LZW.java
This is an Application for the LZW compression technique. it's lossless compression technique. Reading input data (letters or numbers) from the original file (text file) then apply the algorithm steps and push the compressed data on a compressed file. and vice versa, it reads the compressed file and run the decompress function to get the original input data.